replaced type of i to size_t in for-loop

+++ src/jsonConverter.cpp
@@ -81,7 +81,7 @@ void jsonConverter::putAnswers(std::vector<std::vector<relativeIndex>> answers) 
     json j;
     j["answers"];
 
-    for (int i = 0; i < answers.size(); ++i) {
+    for (size_t i = 0; i < answers.size(); ++i) {
         std::string count = std::to_string(i + 1);
 
         for (int k = 3 - count.length(); k > 0; --k) {
